Refactor api calls for single game page
